export useful constants to idl

+++ programs/best-apy/src/instructions/set_protocol_weights.rs
@@ -20,10 +20,10 @@ pub struct SetProtocolWeights<'info> {
     pub vault_account: Box<Account<'info, VaultAccount>>,
 }
 
-pub fn handler(ctx: Context<SetProtocolWeights>, weights: [u32; PROTOCOLS_LEN]) -> Result<()> {
+pub fn handler(ctx: Context<SetProtocolWeights>, weights: [u16; PROTOCOLS_LEN]) -> Result<()> {
     let weights_sum = weights
         .iter()
-        .try_fold(0_u32, |acc, &x| acc.checked_add(x))
+        .try_fold(0_u16, |acc, &x| acc.checked_add(x))
         .ok_or_else(|| error!(ErrorCode::MathOverflow))?;
 
     require!(weights_sum == WEIGHTS_SCALE, ErrorCode::InvalidWeights);
